reorganize sounds, rename some new packages, machine module
