+++ src/main/java/gregtech/api/module/IModuleContainer.java
@@ -1,4 +1,4 @@
-package gregtech.api.modules;
+package gregtech.api.module;
 
 public interface IModuleContainer {
