+++ src/main/java/gregtech/core/CoreModule.java
@@ -12,8 +12,8 @@ import gregtech.api.fluids.MetaFluids;
 import gregtech.api.gui.UIFactory;
 import gregtech.api.items.gui.PlayerInventoryUIFactory;
 import gregtech.api.metatileentity.MetaTileEntityUIFactory;
-import gregtech.api.modules.GregTechModule;
-import gregtech.api.modules.IGregTechModule;
+import gregtech.api.module.GregTechModule;
+import gregtech.api.module.IGregTechModule;
 import gregtech.api.recipes.ModHandler;
 import gregtech.api.recipes.RecipeMap;
 import gregtech.api.recipes.recipeproperties.TemperatureProperty;
@@ -50,7 +50,7 @@ import gregtech.core.sound.GTSoundEvents;
 import gregtech.core.sound.internal.SoundManager;
 import gregtech.integration.theoneprobe.TheOneProbeCompatibility;
 import gregtech.loaders.dungeon.DungeonLootLoader;
-import gregtech.modules.GregTechModules;
+import gregtech.module.GregTechModules;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.world.World;
 import net.minecraftforge.classloading.FMLForgePlugin;
