+++ src/main/java/gregtech/module/GregTechModules.java
@@ -1,11 +1,12 @@
-package gregtech.modules;
+package gregtech.module;
 
-import gregtech.api.modules.IModuleContainer;
+import gregtech.api.module.IModuleContainer;
 
 public class GregTechModules implements IModuleContainer {
 
     public static final String MODULE_CORE = "core";
-    public static final String MODULE_TOOLS = "tools";
+    public static final String MODULE_MACHINE = "machine";
+    public static final String MODULE_TOOL = "tool";
 
     // Integration modules
     public static final String MODULE_JEI = "jei_integration";
