+++ src/main/java/gregtech/tool/ToolModule.java
@@ -1,13 +1,15 @@
-package gregtech.tools;
+package gregtech.tool;
 
 import gregtech.api.GTValues;
-import gregtech.tools.enchants.EnchantmentEnderDamage;
-import gregtech.tools.enchants.EnchantmentHardHammer;
-import gregtech.api.modules.GregTechModule;
-import gregtech.modules.BaseGregTechModule;
-import gregtech.modules.GregTechModules;
+import gregtech.api.module.GregTechModule;
+import gregtech.module.BaseGregTechModule;
+import gregtech.module.GregTechModules;
+import gregtech.tool.enchantment.EnchantmentEnderDamage;
+import gregtech.tool.enchantment.EnchantmentHardHammer;
+import gregtech.tool.sound.ToolSounds;
 import net.minecraft.enchantment.Enchantment;
 import net.minecraftforge.event.RegistryEvent;
+import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -17,12 +19,12 @@ import java.util.Collections;
 import java.util.List;
 
 @GregTechModule(
-        moduleID = GregTechModules.MODULE_TOOLS,
+        moduleID = GregTechModules.MODULE_TOOL,
         containerID = GTValues.MODID,
         name = "GregTech Tools",
-        descriptionKey = "gregtech.modules.tools.description"
+        descriptionKey = "gregtech.modules.tool.description"
 )
-public class ToolsModule extends BaseGregTechModule {
+public class ToolModule extends BaseGregTechModule {
 
     public static final Logger logger = LogManager.getLogger("GregTech Tools");
 
@@ -32,10 +34,15 @@ public class ToolsModule extends BaseGregTechModule {
         return logger;
     }
 
+    @Override
+    public void preInit(FMLPreInitializationEvent event) {
+        ToolSounds.register();
+    }
+
     @Nonnull
     @Override
     public List<Class<?>> getEventBusSubscribers() {
-        return Collections.singletonList(ToolsModule.class);
+        return Collections.singletonList(ToolModule.class);
     }
 
     @SubscribeEvent
