+++ src/main/java/gregtech/tool/enchantment/EnchantmentHardHammer.java
@@ -1,4 +1,4 @@
-package gregtech.tools.enchants;
+package gregtech.tool.enchantment;
 
 import gregtech.api.GTValues;
 import gregtech.api.items.toolitem.ToolClasses;
